feat: component list builder
