+++ src/page/Editor/components/WidgetPickerEditor/components/ComponentPanel/interface.ts
@@ -1,5 +1,6 @@
 import { HTMLAttributes } from "react"
 import { ComponentModel } from "@/wrappedComponents/interface"
+import { SessionType } from "@/wrappedComponents/ComponentListBuilder"
 
 export interface ComponentPanelProps extends HTMLAttributes<HTMLDivElement> {
   componentList?: ComponentSessionProps[]
@@ -8,4 +9,8 @@ export interface ComponentPanelProps extends HTMLAttributes<HTMLDivElement> {
 export type ComponentSessionProps = {
   title: string
   children: ComponentModel[]
+}
+
+export type TypeMapComponent = {
+  [key in SessionType]: ComponentModel[]
 }
