+++ src/wrappedComponents/CanvasWidget/index.tsx
@@ -1,17 +1,20 @@
 import { FC } from "react"
 import { CanvasWidgetProps } from "./interface"
 import { ContainerWidget } from "@/wrappedComponents/ContainerWidget"
+import { SearchIcon } from "@illa-design/icon"
+import { ComponentModel } from "../interface"
 
-export const CANVAS_WIDGET_CONFIG = {
+export const CANVAS_WIDGET_CONFIG: ComponentModel = {
   type: "CANVAS_WIDGET",
   widgetName: "Canvas",
+  version: "0.0.1",
+  icon: <SearchIcon />,
+  sessionType: "COMMON",
   defaults: {
-    version: "0.0.1",
     rows: 0,
     columns: 0,
     width: "100%",
     height: "100%",
-    widgetName: "Canvas",
     containerStyle: "none",
     borderWidth: "0",
     borderRadius: "0",
