+++ src/wrappedComponents/TestWidget/index.tsx
@@ -1,14 +1,17 @@
 import { FC } from "react"
-import { DraggableComponent } from "../DraggableComponent"
 import { TestWidgetProps } from "./interface"
+import { SearchIcon } from "@illa-design/icon"
+import { ComponentModel } from "../interface"
 
-export const TEST_WIDGET_CONFIG = {
+export const TEST_WIDGET_CONFIG: ComponentModel = {
   type: "TEST_WIDGET",
+  widgetName: "test",
+  version: "0.0.1",
+  icon: <SearchIcon />,
+  sessionType: "COMMON",
   defaults: {
     rows: 28,
     columns: 24,
-    widgetName: "test",
-    version: "0.0.1",
   },
 }
