+++ src/wrappedComponents/WidgetBuilder.tsx
@@ -1,50 +1,54 @@
 import { CONTAINER_WIDGET_CONFIG, ContainerWidget } from "./ContainerWidget"
 import { TEST_WIDGET_CONFIG, TestWidget } from "./TestWidget"
 import { CANVAS_WIDGET_CONFIG, CanvasWidget } from "./CanvasWidget"
-import { TEXT_WIDGET_CONFIG, WrappedTextWidget } from "./Text"
-import { IMAGE_WIDGET_CONFIG, WrappedImageWidget } from "./Image"
-import { SWITCH_WIDGET_CONFIG, WrappedSwitchWidget } from "./Switch"
-import { BUTTON_WIDGET_CONFIG, WrappedButtonWidget } from "./Button"
-import { SELECT_WIDGET_CONFIG, WrappedSelectWidget } from "./Select"
+import { TEXT_WIDGET_CONFIG, TextWidget, TextPanelConfig } from "./Text"
+import { IMAGE_WIDGET_CONFIG, ImageWidget, ImagePanelConfig } from "./Image"
+import { SWITCH_WIDGET_CONFIG, SwitchWidget, SwitchPanelConfig } from "./Switch"
+import { BUTTON_WIDGET_CONFIG, ButtonWidget, ButtonPanelConfig } from "./Button"
+import { SELECT_WIDGET_CONFIG, SelectWidget, SelectPanelConfig } from "./Select"
+import { WidgetConfigs } from "./interface"
 
-const WidgetConfig = {
+const WidgetConfig: WidgetConfigs = {
   CONTAINER_WIDGET: {
     widget: ContainerWidget,
     config: CONTAINER_WIDGET_CONFIG,
+    panelConfig: [],
   },
   CANVAS_WIDGET: {
     widget: CanvasWidget,
     config: CANVAS_WIDGET_CONFIG,
+    panelConfig: [],
   },
   TEST_WIDGET: {
     widget: TestWidget,
     config: TEST_WIDGET_CONFIG,
+    panelConfig: [],
   },
   TEXT_WIDGET: {
-    widget: WrappedTextWidget,
+    widget: TextWidget,
     config: TEXT_WIDGET_CONFIG,
+    panelConfig: TextPanelConfig,
   },
   IMAGE_WIDGET: {
-    widget: WrappedImageWidget,
+    widget: ImageWidget,
     config: IMAGE_WIDGET_CONFIG,
+    panelConfig: ImagePanelConfig,
   },
   SWITCH_WIDGET: {
-    widget: WrappedSwitchWidget,
+    widget: SwitchWidget,
     config: SWITCH_WIDGET_CONFIG,
+    panelConfig: SwitchPanelConfig,
   },
   BUTTON_WIDGET: {
-    widget: WrappedButtonWidget,
+    widget: ButtonWidget,
     config: BUTTON_WIDGET_CONFIG,
+    panelConfig: ButtonPanelConfig,
   },
   SELECT_WIDGET: {
-    widget: WrappedSelectWidget,
+    widget: SelectWidget,
     config: SELECT_WIDGET_CONFIG,
+    panelConfig: SelectPanelConfig,
   },
-
-  // BUTTON_WIDGET: {
-  //   widget: ButtonWidget,
-  //   config: BUTTON_WIDGET_CONFIG,
-  // },
 }
 
 export type WidgetType = keyof typeof WidgetConfig
