+++ src/wrappedComponents/interface.tsx
@@ -1,19 +1,27 @@
-import { ReactNode } from "react"
+import { FC, ReactNode } from "react"
+import { SessionType } from "./ComponentListBuilder"
+import { PanelConfig } from "@/page/Editor/components/InspectPanel/interface"
 
 export interface SizeProps {
   w?: number | string
   h?: number | string
 }
 
+export interface WidgetConfigs {
+  [key: string]: {
+    widget: FC<any>
+    config: ComponentModel
+    panelConfig: PanelConfig[]
+  }
+}
+
 export type ComponentModel = {
-  id: string
-  name: string
+  widgetName: string
   icon?: string | ReactNode // url
   type?: string //组件类型
+  version: string
+  sessionType?: SessionType
   defaults?: {
-    version: 1
-    rows: number
-    columns: number
-    // 其他一些默认的属性
+    [key: string]: any
   }
 }
